# SmartEnviMonitoring.API.Test/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SmartEnviMonitoring.API",
    "version": "1.0"
  },
  "paths": {
    "/api/audio": {
      "post": {
        "tags": [
          "Audio"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/device/login": {
      "post": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "deviceUID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/logout": {
      "post": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "deviceUID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/timecurrent": {
      "get": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "deviceUID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/info/logindevices": {
      "get": {
        "tags": [
          "Info"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/info/lastmeasurements": {
      "get": {
        "tags": [
          "Info"
        ],
        "parameters": [
          {
            "name": "num",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReportDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReportDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReportDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/weather": {
      "post": {
        "tags": [
          "Weather"
        ],
        "parameters": [
          {
            "name": "TemperatureC",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Humidity",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "DeviceUID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/{deviceUID}": {
      "get": {
        "tags": [
          "Weather"
        ],
        "parameters": [
          {
            "name": "deviceUID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeasurementRecord"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeasurementRecord"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeasurementRecord"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DeviceDto": {
        "type": "object",
        "properties": {
          "deviceUID": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "lastLoginTimestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DeviceState": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "MeasurementRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "number",
            "format": "double"
          },
          "humidity": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "$ref": "#/components/schemas/MonitoringDevice"
          }
        },
        "additionalProperties": false
      },
      "MonitoringDevice": {
        "required": [
          "deviceUID"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "deviceUID": {
            "minLength": 1,
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "lastLoginTimestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "WeatherReportDto": {
        "type": "object",
        "properties": {
          "temperatureC": {
            "type": "number",
            "format": "double"
          },
          "humidity": {
            "type": "number",
            "format": "double"
          },
          "deviceUID": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}